+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Auto pickup".
 - Tests for "Annihilation".
 - Tests for "Faro shuffle".
+- Tests for "Morellet’s random lines".
 
 ## [1.9.0] - 2022-06-01
 ### Added